+++ src/ts/editorHelpers.ts
@@ -250,12 +250,20 @@ export class EditorHelper {
   }
 
   addNewDialogueNode() {
+    // Place two box units to the right if currently there is a selected node.
+    // Otherwise random position.
+    const position: XYCoords = this.selectedNode
+      ? {
+          x: this.selectedNode.editor.position.x + 2 * this.boxSize.width,
+          y: this.selectedNode.editor.position.y + this.boxSize.height
+        }
+      : this.getRandomPosition();
     const nodeName = this.randomNodeKey();
     const newNode: IMiniQuestionaireWithPosition = {
       q: "",
       o: [{ a: "", next: null }],
       editor: {
-        position: this.getRandomPosition()
+        position: position
       }
     };
     this.dialogConfigWithPositions.graph[nodeName] = newNode;